Rename admin page file
